Adding @since docs for Hyrax::Publisher events
This change adds documentation regarding the specific published events.
I spent a few hours poking around the yard `@!macro` and `@!attribute`
behavior to settle on this.  (Some of the documented behavior doesn't
quite work, and I wonder if its related to the two mixins).

Regardless the goal of this documentation is to help bring in line the
versions in which we started publishing each event.  These `@since`
declarations align with [the corresponding wiki entry][1]

Attached to this commit's associated pull request will be a
screen-capture of how this commit modifies the yard docs for `Hyrax::Publisher`.

[1]:https://github.com/samvera/hyrax/wiki/Hyrax's-Event-Bus-(Hyrax::Publisher)

In addition, I'm adding some emojies to the documentation, as I hope it
draws the eye of future maintainers to say "Hmm, emojiis, that seems
odd, I better check this out."