make Hyrax's user groups configurable
Hyrax has special handling built in for three user groups: admins, registered
users, and the general public. behavior specific to each of these
groups (especially in `Hyrax::Ability` and related mixins) is keyed off the
group name.

in the past, we've provided a kind of configuration at the class level on
`::Ability` (or any application class which mixes in `Hyrax::Ability`), but
centralizing this at Hyrax configuration makes it easier to use across the
application, and easier to understand. it also reduces the risk of inheritance
confusion.

the class level configuration is retained. at least for now, i think it's best
not to deprecate the associated class attributes.