remove dependency injection from WorkflowListener
listeners struggle with dependency injection because they need to be initialized
on app startup. Rails tries to reload their collaborator objects, which leads to
`ArgumentError (A copy of Hyrax::Workflow::WorkflowFactory has been removed
from the module tree but is still active!)`

pending a better fix, we just need to ban DI for listeners and hard code their
collaborators.