group and document `ability_logic` hook methods in Hyrax::Ability
many of the methods in `Hyrax::Ability` are utilities to be included in (or
dropped from) `.ability_logic`. these methods are `private` at the Ruby
level (i.e. they should only be called from this `Ability` instance), but are
public API (since they are intended for use in applications).

we take a stab at documenting them here.