add a rake task to update collection references to collection types
as part of #4696: since we're deprecating our older semi-custom global id
syntax, we need to be able to support applications updating their collections.

the rake task `hyrax:collections:update_collection_type_global_ids` provides a
way to do this, and is a non-destructive if collections all use the standard
syntax.

note that this task may create substantial load, even if all collections are
already up to date; so it shouldn't be run repeatedly.