implement a generic query for finding ids by model
this makes it possible to query Valkyrie adapters to filter a set of ids down to
those matching a model. if no ids are passed, all ids matching the model are
returned.

this default implementation follows our policy of providing an implementation of
all custom queries that works for Hyrax using only Valkyrie's built-in
queries. follow-up work will implement for Wings.

related to #4722.