use a custom query to find ids matching a model
the inline implementation of this query logic was repeated in multiple places,
and didn't scale (if there were many collections associated with the user via
permission templates, the query would pass too many arguments to Solr). the
query is now supported in Hyrax as a Valkyrie custom query, so we can use that
and trust the adapter to provide an efficient implementation.