reindex ActiveFedora objects when valkyrie ACLs are updated
this is a bit tricky...

Valkyrie models treat ACLs as separate, first class resources that refer to the
resource they grant access on. this means, a work/file_set, etc... must exist in
order to create an ACL and grant permissions. in our Valkyrie usage, we don't
normally want to set ACLs directly via the object, we work through
`Hyrax::AccessControl` and the service-level tooling `Hyrax::AccessControlList`
instead.

so, normally, we create a Resource, then update the ACLs and the resource
independently. `Blacklight::AccessControls` and  `Hyrax::Ability` generally
count on the index being up-to-date, so we have a listener to watch for ACL
updates and update the index.

the tricky bit is that with Wings, our indexer is generally a no-op, since
updating indexes is handled as part of the persister (ActiveFedora::Base#save
does both metadata persistence and index updating). calling `#save` via
`ActiveFedora` does in fact update the index's ACL data to the current
state. but we can update the ACL at any time, and we don't necessarily update
the resource at the same time!

this is especially an issue on object creation: we can't persist the ACL before
we create the object. we could add a bunch of special handling to Wings to make
sure ACLs are added in the way `Hydra::AccessControls` wants. but this special
handling makes ensuring that we get the end state we want a bit complicated. so
the propasal here is to avoid special handlind and instead introduce a special
listener, just for Wings, that updates the (ActiveFedora) object index whenever
the ACL is up-to-date.

this encourages us to follow the new pattern, and code written to play nice with
this should also play nice when the Valkyrie indexer isn't a no-op.