refactor AdminAdminSetMemberSearchBuilder specs
there was a bunch going on in this test setup that wasn't really used. there
also seems to be more going on in this builder than is really justified as well,
especially for a class that's only used in one place.

ripping up the specs then moving closer to the controller to see what's up.