remove duplicate `#gated_discovery_filters` implementation
this implementation exactly duplicates the inherited implementation from the
parent class. remove it.

this is a first step to completely removing this filter strategy. the whole
exercise here is to force Solr to return us documents for collection ids based
on criteria that are stored in the database (permission template), not the
index. there's just no good reason to do this other than code structure
convenience, so we should rework the whole thing.