Aligning batch size and page size enforcement
Closes #3242

Prior to this commit, there existed a ["magic number" in the
Hyrax::MyControllers][1].  This "magic number" was less than the largest
value in [the range for
`app/views/hyrax/my/_sort_and_per_page.html.erb`][2].

When a user selected a page size greater than the "magic number", then
they would see [a CSS overlay][3]. The CSS overlay is an 6 year old UI
implementation that provides a little bit of guidance were someone to
manufacture a URL with a custom `per_page` parameter.

Wit this commit, I'm aligning the "magic number" with the UI
constraint, and exposing a configuration option.  Enforcing the maximum
is likely a good idea to prevent (or maybe just discourage) custom
`per_page` parameters from over working the batch system.

[1]:https://github.com/samvera/hyrax/blob/0222988b13ddf03cb584c8c9168da7c8d826b674/app/controllers/hyrax/my_controller.rb#L56
[2]:https://github.com/samvera/hyrax/blob/0222988b13ddf03cb584c8c9168da7c8d826b674/app/views/hyrax/my/_sort_and_per_page.html.erb#L8
[3]:https://github.com/samvera/hyrax/blob/0222988b13ddf03cb584c8c9168da7c8d826b674/app/assets/stylesheets/hyrax/_file-listing.scss#L157