Conditionally reseting fileset depositor for AF
Starting with the [change_content_depositor_service_spec.rb][1], we have
a test that asserts the current behavior.

(awead) [introduce the spec in a 2015 commit][2], I found this commit by (grosscol) [that subtly removed the expected behavior][3]

So, my analysis is that one commit introduced the bug and a later spec
asserted the bug as expected behavior.

This commit addresses the ActiveFedora bug.  I add this as a separate
commit, because I'm about to add a spec for Valkyrie that will
structurally change spec I updated in this commit.

Ref #3260

[1]:https://github.com/samvera/hyrax/blob/508e7381d9b5cfce4cd27c40406141df47876eca/spec/services/hyrax/change_content_depositor_service_spec.rb#L34:L49
[2]:https://github.com/samvera/hyrax/commit/468929845f2989913617a297af801bbda6bd8e45
[3]:https://github.com/samvera/hyrax/commit/0cb55bf46092bbb72c1803e7341074e1ea903a5b#diff-bfac0ce58388b8031096668717616b32e0edbeaf3707ca4cde388ee859dcdd35L20-L24