Adding title attribute for visibility link
Given that we're linking to the "share" tab, I opted to include the I18n
translation of the [Share tab][#1].  I'm including it as a prefix for
the title.  To do much else would require another translation entry.  I
believe this is adequate to provide enough differentiation in the titles.

Related to #4753

[1]:https://github.com/samvera/hyrax/blob/001bb7ad73300a2183ed681c0b4e917f25199b67/app/views/hyrax/base/_guts4form.html.erb#L21