Adding `aria-hidden=true` for thumbnail tag
Prior to this commit, there existed two URLs that linked to the same
resource.  One A tag wrapped a thumbnail, the other A tag wrapped the
title of the work.

From an accessibility stand-point, there need only be one link to the
document (two links to the same URL can be confounding).

By setting aria-hidden, we're telling assistive devices that this link
isn't of importance.  We include the link as a courtesy for those
relying on the non-assistive rendering.

Related to #4753