add hook for avoiding Google Scholar embed for non-scholarly content
Google Scholar only wants "scholarly articles". many objects deposited in
Hyrax apps are not that. it should be possible for applications to easily
determine which content should get google scholar embeds and which should
not. Using a presenter/decorator gives us a place to determine this without
adding to existing model or presenter APIs.

We'll reuse this presenter later to handle the value presentation issues
currently handled by view logic.