refactor file manager form tests, extend MemberPresenterFactory
use named subject; avoid indirection in test logic. these tests don't exercise
any of the real behavior of this module; from their perspective, it's not much
more than a wrapper for the member factory. since the tests want to inject a
PresenterFactory, we can guess that others might too (will we need custom ones
for valkyrie?), so we extend the class to support factory injection.