allow helm users to opt out of uploading configs to solr on deploy
there's a handy set of scripts wrapped in an init container that allows the
chart to maintain the solr collection configuration on deploy.

these scripts make some assumptions that might not be true in all contexts; most
importantly that authentication is turned on for the solr cluster. giving users
the ability to opt out if they want to setup their configset and collection
manually is helpful.