configure a FCREPO_BASE_PATH for helm deploys
since it may be desirable to deploy multiple hyrax instances on a single
FCRepo, we use a name more unique that `/prod` for the Hyrax base path. using
the Helm release name is useful, but will run into trouble if multiple releases
with the same name are pushed to different namespaces or different clusters. for
this case, we provide an `fcrepoBasePathOverride` to allow users to choose
something suitably unique.