support external postgresql in helm chart
enables use of external/existg postgresql.

this will fail if the database doesn't already exist. when we use an internal
postgresql subchart, we count on the chart arguments/initialization to create
the database. we can't count on that here. so, in some sense, this is partial
work. on the other hand, we can wave our hands a bit and say that if you're
going to stand up your own postgres you can create the database while you're at
it.

maybe some follow-up work to allow an `ensure-db` init pod is called for?