add a db:create step to the helm deployment process
ensures that the configured database exists by adding a `rake db:create`
invocation to the `db-setup` initContainer.

in the event we're trying to deploy hyrax against a database engine that it has
permissions for, but where the configured database doesn't exist, this will
ensure it gets added before we try to run the migrations.