extend FileSetAttachedEventJob to work with Valkyrie
support valkyrie style lookup of parent works for the event job. there's good
reason to think this job could be quicly converted to valkyrie-only by removing
the switch statement and possibly using an id-based inverse relation lookup for
membership instead of `find_parents`. for now, it seemed best to leave the
behavior for ActiveFedora totally fixed.