wings: fix serveral shared example specs for Valkyrie persister
this reduces the space of shared example failures to type issues, order
issues, and nesting issues. 22 tests fail in total.

RDF.rb and ActiveFedora cast many RDF types to Ruby objects, which is mostly
helpful, and we don't have plans to circumvent this to pass Valkyrie tests.

ordered attribute support is forthcoming, and nesting is to be explored.