chart: setup hyrax-worker image for Helm chart
use a separate pod to run hyrax workers.

i think the base configuration of this deployment should be roughly okay with
whatever image or ActiveJob adapter might be deployed to run the background
jobs. as long as the image's command will run the background jobs and will
shutdown it it won't service new jobs

the chart's branches could probably be better organized to avoid sidekiq cruft
when sidekiq is not in use, but first i want to work on getting volumes setup
reasonably.