don't break on UrlHelper load if models aren't ActiveModel conformant
not all possible `registered_curation_concern_types` (or `FileSet`s or
`Collection`s) have a model name. likewise, `registered_curation_concern_types`
contains strings, which might not be constantizable by mistake. if the rest of
the application works under these conditions, just skip defining track paths.

note that this module is eagerly loaded and run (only in production, for usual
app configurations!) even if it's not included anywhere in the downstream
application. this is a bad time to break someone's app on startup over code they
didn't load themselves.