chart: add environment variables/secret for the db username and pass
in the case that an application wants to connect to the postgresql instance
outside the database configured with `DATABASE_URL` it's handy to have these
values. all this information is already exposed to the application via the
`DATABASE_URL` secret entry, so passing this along shouldn't have any security
implications.