Support dynamic initContainers in Helm chart
For both the main application deployment as well as the worker
deployment, this adds the capability for downstream chart users to
include their own initContainers.

A use case for this might be waiting on an additional database not
natively monitored by the chart, but is required to be up and running
for the application to start.