chart: use OCI published chart for fcrepo
we've been getting 404 errors from the fcrepo chart, causing trouble on
builds. since Samvera is now publishing charts to the GitHub OCI registry, maybe
we want to just use that?

the tricky thing is that OCI support is "experimental" and requires
`HELM_EXPERIMENTAL_OCI=1` to be set in the environment.