allow the SolrCloud assign-configset utility to create collections
many solr images provide an easy way to create an inital collection via the
entrypoint, so this script was originally written with the assumption that the
collection would already exist.

in practice, it's nice to be able to create multiple collections (e.g. to add a
test collection) or to create one that doesn't exist (e.g. because the solrcloud
instance is preexisting and has other collections, but not ours). this work
extends the script to allow creating the collection in the init container if it
doesn't already exist.

this also fixes an issue where the init container would loop and timeout when
the collection is missing. it will now try to create the collection instead.