Ensuring non-parseable date bubbles up to UI
Prior to this commit, we were not verifying that a user's input for a
release_date was in fact a parseable date.  Chrome and Firefox apply a
date picker "widget" that enforces entries for a `type="date"`.
However, Safari does not.

Closes #4757

Note, to test this in the UI requires a fix for #4833.  That fix, at the
time of crafting this commit message, can be seen at PR #4885.