Ensuring admins can manage participants
Based on the descriptive text in the [_form_participants][3] (see
below), the application describes that a manager of an admin set should
have the rights to manage the participants of objects deposited into the
admin set.

> Managers of this administrative set can edit the set metadata,
> participants, and release and visibility settings. Managers can also
> edit work metadata, add to or remove files from a work, and add new
> works to the set.

The implemented solution favors an initial check that the given ability
can manage the associate permission template.  The admin set form
partials ([panel navigation][1], [renderinig panel content][2], and
[panel content][3]) support this short-circuit.  A more conservative
approach would be to first check that we have an active workflow, then
test the ability, and finally if the workflow allows for access
granting.

I chose to go with the path that extends more privileges to the
administrator.

Closes #3175

[1]:https://github.com/samvera/hyrax/blob/a8e1d84e48d7444d4f5dce4ada7a82dd1dad7737/app/views/hyrax/admin/admin_sets/_form.html.erb#L43
[2]:https://github.com/samvera/hyrax/blob/a8e1d84e48d7444d4f5dce4ada7a82dd1dad7737/app/views/hyrax/admin/admin_sets/_form.html.erb#L8-L10
[3]:https://github.com/samvera/hyrax/blob/a8e1d84e48d7444d4f5dce4ada7a82dd1dad7737/app/views/hyrax/admin/admin_sets/_form_participants.html.erb#L1