Echoing tab rendering selection logic
Given that later logic in this view partial has `(tabs - ['share'])` it
makes sense to echo that concern here.  Also, given that we are
explicitly including a share navigation LI, it would be confusing to
have two 'share' elements show up.  This change aligns the rendering
logic of both the navigation and content.