MemberPresenterFactory needs tests
this is a complicated class, but only one method is tested. even that test is
thin and heavily stubbed. let's work on getting some real tests in here so we
can support Valkyrie.

the rubocop disable addition is because the existing stubs make rubocop
mad (yes, they are very bad!) when it knows they refer to the unit under
test (`subject(:factory)`).