add a MemberPresenterFactory implementation that supports valkyrie
the existing `MemberPresenterFactory` queries for members in solr, which seems
appropriate. even with a fast backing database, much of our Blacklight frontend
behavior is going to be based on the Solr index. Since the various Presenters
consume `::SolrDocument`s, keeping it all in Solr seems best.

however, `MemberPresenterFactory` is also unduly complicated, queries solr many
times to produce one list of members, and keys off of ActiveFedora-specfic Solr
keys. it does this to retain order, and to provide separation of FileSet and
Work presenters.

the latter is dealt with here by using the indexed "generic type". the former is
still WIP; we can reorder the documents to match the `member_ids` (which are
ordered in Valkyrie).

this should also be extended to provide filtering by permissions at query time.