Verifying round trip AF conversion validation
The desired behavior is as follows:

Given `x` is an ActiveFedora object which responds to `x.valid?` and returns `false`
And `f(x)` converts `x` to a `Valkyrie::Resource` (which doesn't respond to `valid?`)
And `g(f(x))` converts `f(x)` resource back to an ActiveFedora object that responds to `valid?`
Then `g(f(x))` responds to `valid?` and returns `false`

This spec tests that behavior.  However, I'm not able to duplicate the
error.  Perhaps it's a matter of the specific parameters?  I'm adding
this spec as a placeholder for what could be the specific
attributes/values that trigger a failing scenario.

Closes #3917 as I'm unable to duplicate the error