Removing excessive mocking to better test
Given the labyrinth that is the implementation of the FileSet jobs and
actors, it seems quite reasonable to remove the mocks and instead rely
on a more full stack test.

This also helps us know that when we use an ActiveFedora object it
works and by extension when we add tests/behavior for
Valkyrie::Resource objects, we'll know it works as well.