Ensuring IngestLocalFileJob handles valkyrie obj
This commit documents and tests that the IngestLocalFileJob handles
valkyrie resources.  We wouldn't have needed to make any changes to this
job to verify that Valkyrie worked.  However, by favoring an explicit
change a future maintainer can quickly look at this class and say "Cool,
I don't need to check if this works with Valkyrie!"

Closes #4196