Renaming variable for better clarity
Prior to this commit, when I was reading the code I saw "f" and
"file_set" in the same context.  That "f" is then passed to another
method which clarifies what it's type.

This commit looks to separate the "f" (which implies "file") from the
"FileSet" which is the container of a "file".