refactor PermisionsController to use `load_resource`
use CanCanCan's `load_resource` to resolve the resource in
`PermissionsController`. this has the advantage of centralizing the choice of
model class and persistence service to use.

to close #4917, we'll need to ensure the background jobs support Valkyrie, and
then switch to using `Hyrax::Resource` as the model to resolve.

in specs: nest legacy AF models. remove some stubs in favor of real
actions. rework ActiveJob specs. drop specs for deprecated controller actions.