deprecate `Suppressible#suppressed?`
this behavior has been extracted to `Hyrax::ResourceStatus.new(resource:
self).inactive?`. since it won't be used on models going forward it's important
to stop calling it. since we've encouraged folks to override this method with
custom behavior, we can't stop calling it from some key places in Hyrax yet,
mainly the indexers.

`#suppressed?` is still used on `SolrDocument`, and we'll continue to index
`suppressed_bsi` and provide `#suppressed?` for `SolrDocumentBehavior` and
presenters.