make VersionPresenterList an Enumerable; extract creation to factory
`VersionPresenterList` is effectively an `Enumerable` (responds only to
`#each`), extending it with `include Enumerable` makes it more
versitile (perhaps we should consider a rename?). the current implementation
relies on loading each member up-front and then sorting, but this might not be
necessary for all backends. `Enumerable` also provides some guidance about how
to do this kind of loading lazily.

move the generation to a factory so the details of how to resolve an object to a
list of versions are encapsulated.