extract some `FileSetsController` behavior into form
we want to support this behavior across data middlewares. extracting it out of
the controller feels like the best first step.

the `FileSetsController` and its related views do a lot of assignments, which
spreads their ActiveFedora dependencies out quite a bit. the idea here is to
consolidate the behavior in the form(s), then move the views to rely on the
forms. the end result will be that the dependency footprint for the
controller-view complex will be reduced.

we keep the variable assignments to support custom views (it's a common pattern
to override and customize behavior directly in views). they'll need to be
removed on a major version bump.