extract view dependency on NestedCollectionQueryService to helper
the subcollection modal had a hard coded dependency on
`NestedCollectionQueryService`, which depends on very specific indexing behavior
for collections. extracting that dependency to a helper makes it easier to
control the specific behavior.

in writing tests for this new helper, i tried to analyze the behavior of the
downstream method. the current tests there are heavily stubbed, and it's not
clear which collections are *supposed* to be included in this list (apparently
something to do with the collection type?). the current tests are therefore
minimal, and test the current behavior.