don't reload the collection to populate subcollection modal
this view used to call `#find` to retirieve a collection from an id. we already
have a presenter and a solr document, so there's no need for this.

instead, allow `Hyrax::Ability` to resolve `CollectionPresenter` to its
underlying `SolrDocument`. Add `CollectionPresenter#nestable?` as an alias to
the existing `#collection_type_is_nestable?` to match the `Collection`
interface, and pass the presenter through to the modal. we continue to pass the
id into the modal for backward compatibility.