don't show the "Add to collection" action when type isn't nestable
the previous guard skipped showing this button only if no collection types were
nestable (checking this involved a solr query). it would then base the actual
behavior on the current collection type's nestability. in the case of admin
sets, it was even hard coded to false!

instead, check whether the current type is nestable. this avoids the most common
case where the user might push this button only to be given a deny popup. since
none of the data is loaded in dynamically in response to the button push,
there's no reason we can't remove that modal implementation altogether.