move parent collection JSON generator into helper
the implementation of query for eligible parent collections had a handful of
problems.

first, it was awkwardly located in the presentation layer. since this code is
really about a business rule ("which parent collections are eligible for this
object?"), hiding it in presenters makes it hard to understand how to approach
changes. it was also duplicated across `CollectionForm` and
`CollectionPresenter`, compounding this problem. moving it to a helper doesn't
completely solve this problem, but it does give us a single point of focus for
continued refactors. as it stands, the business rule is tightly coupled to both
the JSON data generation and the underlying query. we'll probably want to loosen
that up before changing any of these three separate concerns.

second, the implementations retrieved a full object from Fedora. this is
particularly worrying because of how many collections are likely to be included
in a page. in general, we should avoid database access initiated by a
view (controllers should do this!). in the particular case, the called code
doesn't actually need the full object so we can just skip the query. there's
still a view-initiated solr query here, but baby steps.

lastly, there were some bugs in the specific existing implementations: they
cached and returned the wrong data type for repeated calls (Array<Hash>, vs JSON
String). i think this caching is completely unused, so i dropped it from the
helper's implementation.