wings: versions for fcrepo
support a versions interface for the valkyrie storage adapter. the interface
developed here only applies to the access side---not to explicit version
creation. since `fcrepo` creates the versions for us in the normal
configuration, we can just rely on that for this adapter.

further work is needed to validate that this interface is sufficient for Hyrax,
and that some version of it can be implemented for other versioning storage
layers (e.g. S3).