Updating kubernetes README
These were the instructions that I followed to stand up a kubernetes
cluster.

Note, in the case of the ImagePullBackOff error, I tried clearing my
image cache and all kinds of "start with a clean slate."  However, the
errors continued to be somewhat erratic.  I was running through an 8MB
up/1MB down DSL connection and kept encountering problems.  I moved to a
15 MB up/15 MB down fiber connection at another location and the
instructions worked without a hitch.