chart: re-instate external fcrepo hostname support
when the fcrepo chart is disabled, still set the `FCREPO_HOST` and
`FCREPO_BASE_PATH` and other environment variables. the templates and variable
logic should (and already do?) handle all cases.

i think this regressed because of UC work to make it possible to simply skip
fcrepo. setting the environment variables is harmless in this case though. i
think we just got a bit overzealous.