Adding documention on located_near attribute
As I'm working through #1065, I needed to understand the default
behavior.  I also wanted to look and see if there were more attributes
on the RDF document that I could add to the label.

I'm also considering that #1956 required a change in implementation;
moving away from RDF and towards the JSON response document.

At present, I'm uncertain on the best approach to proceed with resolving
issue #1065.  I'll be bringing this up in the hyrax-wg channel and
during stand up.