refactor AdminSetCreateService to use valkyrie throughout
this service still accepts ActiveFedora objects, but casts them to valkyrie and
uses the newer models to handle the updates.

there's still a gap here related to PermissionTemplate ACL setting, which
probably needs to be extracted into another service. the logic for this isn't
immediately clear to me, and the dependencies are a bit of a nest.

commiting this as substantial WIP, and setting this class aside to untangle the
PermissionTemplate stuff.