add top level module documentation for Sipity with some use cases
this is an attempt to document Sipity at the top level. the focus here is on
describing the domain model and explaining some of the use cases. i think a lot
more is required to make this crystal clear; e.g. example code, state diagrams,
and a deep description of Roles, Notifications and Methods is probably called
for.