Split off hyrax-work-base image and bundle install from dassie (#4997)
* Split off hyrax-work-base image and bundle install from dassie

* Skip including the dassie Gemfile when building in docker-compose

* Fix bundler dependency errors

The primary cause of the dependency errors seems to be a combination of newer versions of bundler (2.2+?) and use of a scoped rubygems source without a global source. Having locally cached gem metadata from previously installed gems could  prevent the failure, but a clean environment, such as a docker image build, is particularly vulnerable. Dassie now attempts to only add the global source once.

Some errors were being silently caught as part of a hack to allow the hyrax gem to be required twice. This is fixed by using gemspec with the path option so the bundler resolver sees these as identical.

Gemfile.lock files have been added to dockerignore to prevent the local environment from contaminating the image.

Unpredictable behaviour with which version of bundler is used during image builds was observed. Old versions could persist in the hyrax-base image. In the dev images, bundler is now updated prior to installing gems, and then old versions cleaned.

Co-authored-by: tamsin johnson <tomjohnson@ucsb.edu>