add an the AdminSetOptionsPresenter alternative that avoids queries
`AdminSetOptionsPersenter` does extensive querying: first of admin sets, and
their permissions then of `PermissionTemplates` and lastly of the
`#active_workflow`. in the existing architecture, this is all initiated by view
code (recently, it has moved to a helper). ideally the Controller should broker
the querying and simply provide the relevant data to the view.

in order to achive this, we need a new presenter that accepts the needed input.
`AdminSetSelectionPresenter` in designed to fill this role and eventually
replace `AdminSetOptionsPresenter` completely. it accepts an enum of
`admin_sets` on initialization, and uses them to build the underlying data. the
plan from here is:

  - extend `AdminSetSelectionPresenter::OptionsEntry` to accept a
    `PermissionTemplate` and generate `#data` according to its settings.
  - refactor the `admin_set_options` helper to use a prepopulated presenter if
    provided by the controller.
  - implement the controller behavior to query the data eagerly.
  - optimize queries.