Set user to root for sidekiq compose service
I was having issues using the compose development environment locally. I
believe the issues I experienced were along the lines of:

```sh
sidekiq_1     | There was an error while trying to write to
sidekiq_1     | `/app/samvera/hyrax-webapp/Gemfile.lock`. It is likely that you need to grant
sidekiq_1     | write permissions for that path.
```

And I believe this is due to the `sidekiq` service needing to run as
`root` like the `app` and `db_migrate` services.