refactor Workflows::StatusListService to accept a user directly
this service had a highly bidirectional dependency on its calling
controller. this kind of dependency makes testing harder, and limits reuse. in
practice, all this service really needs is a `::User` and the set of filters for
the workflow query (can this be abstracted?). refactor to allow the service to
accept just a user instead of a search context/controller, and deprecate the
controller arugemnt.

moving toward valkyrie support for `Workflows::StatusListService`.