fix solr query in MultipleMembershipChecker  and refactor to improve readability
The code was complex and difficult to follow.  I refactored so that each method has one and only one task and is clearly named for that task.

Also, tests were heavily mocked and were hiding errors in the generated solr query.  The tests are rewritten to organize them into clear test scenarios with real solr access.