add a rake task to regenerate derivatives
it's a common pitfall in deploying Hyrax apps to misconfigure
thumbnail/derivatives path. when the configuration changes, the thumbnails will
be missing. depending on whether the storage medium they were originally
generated to still exists, it can be impossible to recover the old thumbnails.

adding a task to regen the derivatives completely seems of high utility.