Run extraInitContainers after chart initContainers
Downstream extraInitContainers may be dependent, in a way, on the
upstream initContainers running first.

For example, if one wants to add an additional Solr collection (say for
running tests against). If a local `extraInitContainer` is created to do
this, it may be implicitly dependent on the `load-solr-config`
`initContainer` having already ran (to reference the `configset`).