fix an unreliable test in ActionableObjects specs
this test only worked if the Sipity::Roles and Sipity::WorkflowRoles are 1:1,
which they aren't if another workflow exists with different WorkflowRole
definitions/ids. using the Sipity::WorkflowRole id instead of the one for
Sipity::Role guarantees the correct result.

with this fix in, there's no longer a need to clean the repo between runs.